requirements.txt
aiofiles==0.7.0
altgraph==0.17
amqp==5.0.6
anyio==3.3.0
asgi-csrf==0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-default-group==1.2.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
datasette==0.58.1
datasette-vega==0.6.2
Django==3.2.6
django-environ==0.4.5
dukpy==0.2.3
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
hupper==1.10.3
idna==3.2
itsdangerous==2.0.1
janus==0.6.1
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
mergedeep==1.3.4
packaging==21.0
Pint==0.17
pluggy==0.13.1
prompt-toolkit==3.0.19
pyinstaller==4.5
pyinstaller-hooks-contrib==2021.2
pyparsing==2.4.7
python-baseconv==1.2.2
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
rfc3986==1.5.0
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
ultimate-sitemap-parser==0.5
urllib3==1.26.6
uvicorn==0.14.0
vine==5.0.0
wcwidth==0.2.5
